pyproject.toml
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"


[tool.poetry]
name = "useless-discord-bot"
version = "0.1.0"
description = ""
authors = ["Parnassius <Parnassius@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "3.9.*"
environs = "9.3.2"
"discord.py" = "1.7.2"

[tool.poetry.dev-dependencies]
black = "21.5b1"
isort = "5.8.0"
mypy = "0.812"
pylint = "2.8.2"


[tool.pylint.master]
load-plugins = [
    "pylint.extensions.bad_builtin",
    "pylint.extensions.check_elif",
    "pylint.extensions.typing",
]

[tool.pylint.messages_control]
disable = [
    "invalid-name",
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
]

[tool.pylint.format]
max-line-length = "88"

[tool.pylint.typing]
py-version = "3.9"
runtime-typing = "yes"
